pyproject.toml
[project]
name = "last-chance-at-life"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "groq>=0.30.0",
    "matplotlib>=3.10.3",
    "mistralai>=1.9.2",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "textblob>=0.19.0",
    "cerebras-cloud-sdk>=1.23",
]
